@@ -10,16 +10,17 @@
     "dependencies": {
         "axios": "^0.21.1",
         "core-js": "^3.6.5",
-        "vue": "^2.6.11"
+        "vue": "^3.0.0",
+        "vue3-click-away": "^1.2.1"
     },
     "devDependencies": {
         "@vue/cli-plugin-babel": "~4.5.0",
         "@vue/cli-plugin-eslint": "~4.5.0",
         "@vue/cli-service": "~4.5.0",
+        "@vue/compiler-sfc": "^3.0.0",
         "babel-eslint": "^10.1.0",
         "eslint": "^6.7.2",
-        "eslint-plugin-vue": "^6.2.2",
-        "vue-template-compiler": "^2.6.11"
+        "eslint-plugin-vue": "^7.0.0"
     },
     "eslintConfig": {
         "root": true,
@@ -27,7 +28,7 @@
             "node": true
         },
         "extends": [
-            "plugin:vue/essential",
+            "plugin:vue/vue3-essential",
             "eslint:recommended"
         ],
         "parserOptions": {
